+++ src/Views/src/Compilers/ICompiler.php
@@ -12,9 +12,7 @@ declare(strict_types=1);
 
 namespace Opulence\Views\Compilers;
 
-use Exception;
 use Opulence\Views\IView;
-use Throwable;
 
 /**
  * Defines the interface for view compilers to implement
@@ -26,7 +24,7 @@ interface ICompiler
      *
      * @param IView $view The view to render
      * @return string The compiled view
-     * @throws Exception|Throwable Thrown if there was an error compiling the view
+     * @throws ViewCompilerException Thrown if there was an error compiling the view
      */
     public function compile(IView $view): string;
 }